refactor: move aws methods to submodule
